\name{cluster.q}
\alias{cluster.q}

\title{Clustering Goodness measured by Q2}

\description{Calculates a goodness of clustering measure based prediction sum squares.}

\usage{cluster.q(data,cluster)}

\arguments{
\item{data}{The data matrix}
\item{cluster}{a vector descibing the cluster memberships}
}

\author{Per Broberg}

\value{The clustering mean Q2}

\references{
Eriksson, L., Johansson, E., Kettaneh-Wold, N. and Wold, S. (1999)\emph{ Introduction to
Multi- and Megavariate Data Analysis using Projection Methods (PCA \& PLS)}, Umetrics
}
\keyword{multivariate}
