\name{mat2TeX}

\alias{mat2TeX}

\title{Ouput matrix to LaTeX}

\description{The function outputs a matrix to a LaTeX table}

\usage{mat2TeX(mat, digits = 4, rowNameTitle = "", file = "",
 roundNum = NULL, rowNameAlign = "l", matAlign = "r",
 prtHead = TRUE, prtEnd = TRUE, extraTitle = NULL,
 rowNameCols = 1, append = FALSE)}

\arguments{
\item{mat}{a matrix}
\item{digits}{number of digits}
\item{rowNameTitle}{title above row names}
\item{file}{output file}
\item{roundNum}{}
\item{rowNameAlign}{alignment of row names}
\item{matAlign}{}
\item{prtHead}{}
\item{prtEnd}{}
\item{extraTitle}{}
\item{rowNameCols}{}
\item{append}{}


}

\author{Juerg Kindermann; code found on R list}

\keyword{IO}
